restructured the gui package
