+++ health/src/main/java/com/health/gui/input/VInputPanel.java
@@ -1,8 +1,8 @@
-package com.health.gui;
+package com.health.gui.input;
 
 import javax.swing.JPanel;
 
-import com.health.gui.fileSelection.FileSelectionPanel;
+import com.health.gui.VidneyPanel;
 import com.health.gui.xmlwizard.XmlWizard;
 
 /**
