+++ health/src/main/java/com/health/gui/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Package documentation goes here.
+ * Package containing the GUI classes.
  */
 package com.health.gui;
