+++ health/src/main/java/com/health/gui/script/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author Bjorn van der Laan
+ *
+ */
+package com.health.gui.script;
